onboarding_steps: Validate parameters with Zod.
