+++ web/src/onboarding_steps.ts
@@ -1,19 +1,8 @@
-import {z} from "zod";
+import type {z} from "zod";
 
 import * as blueslip from "./blueslip";
 import * as channel from "./channel";
-
-const one_time_notice_schema = z.object({
-    name: z.string(),
-    type: z.literal("one_time_notice"),
-});
-
-/* We may introduce onboarding step of types other than 'one time notice'
-in future. Earlier, we had 'hotspot' and 'one time notice' as the two
-types. We can simply do:
-const onboarding_step_schema = z.union([one_time_notice_schema, other_type_schema]);
-to avoid major refactoring when new type is introduced in the future. */
-const onboarding_step_schema = one_time_notice_schema;
+import type {StateData, onboarding_step_schema} from "./state_data";
 
 export type OnboardingStep = z.output<typeof onboarding_step_schema>;
 
@@ -45,6 +34,6 @@ export function update_onboarding_steps_to_display(onboarding_steps: OnboardingS
     }
 }
 
-export function initialize(params: {onboarding_steps: OnboardingStep[]}): void {
+export function initialize(params: StateData["onboarding_steps"]): void {
     update_onboarding_steps_to_display(params.onboarding_steps);
 }
